requirements.txt
fastapi>=0.68.0
uvicorn
pydantic==1.10.13
python-dotenv
requests
slowapi
python-multipart
sqlalchemy
psycopg2-binary
pandas 
pillow  # For image handling